# contracts/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Share Disk Control API
  description: |
    Control plane API for the personal distributed cloud drive system.
    This API handles authentication, device management, file catalog, transfers, and synchronization.
  version: 0.1.0
  contact:
    name: Share Disk Team
  license:
    name: MIT
servers:
  - url: http://localhost:8080
    description: Local development server
security:
  - bearerAuth: []
paths:
  /v1/browser/login:
    post:
      summary: Create an opaque browser session without registering a storage device
      operationId: browserLogin
      security: []
      tags: [auth]
      description: Requires an exact same-origin Origin header. Sets an HttpOnly SameSite=Strict cookie. TLS origins receive Secure cookies. Credentials are never returned in JSON.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [account, password]
              properties:
                account: { type: string }
                password: { type: string, format: password }
                remember: { type: boolean, default: false }
      responses:
        '200':
          description: Browser session established
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BrowserSession' }
        '401': { description: Invalid credentials }
        '403': { description: Origin rejected }
  /v1/browser/session:
    get:
      summary: Restore the current browser session and obtain its CSRF token
      operationId: browserSession
      security: [{ browserSession: [] }]
      tags: [auth]
      responses:
        '200':
          description: Current session
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BrowserSession' }
        '401': { description: Session absent or expired }
  /v1/browser/logout:
    post:
      summary: Revoke the browser session and expire its cookie
      operationId: browserLogout
      security: [{ browserSession: [] }]
      parameters:
        - $ref: '#/components/parameters/BrowserCSRF'
      tags: [auth]
      responses:
        '204': { description: Session revoked }
        '403': { description: Origin or CSRF rejected }
  /v1/browser/files/{fileID}/download-ticket:
    post:
      summary: Issue a one-minute file and origin scoped direct Agent download ticket
      operationId: browserDownloadTicket
      security: [{ browserSession: [] }]
      tags: [catalog]
      parameters:
        - { name: fileID, in: path, required: true, schema: { type: string, format: uuid } }
        - $ref: '#/components/parameters/BrowserCSRF'
      responses:
        '200':
          description: Submit this ticket in a form POST directly to the returned Agent URL
          content:
            application/json:
              schema:
                type: object
                required: [url, ticket, expires_in]
                properties:
                  url: { type: string, format: uri }
                  ticket: { type: string }
                  expires_in: { type: integer, const: 60 }
        '409': { description: No online replica }
  /v1/lan/browser-download:
    post:
      summary: Stream a verified Agent file as a browser attachment
      operationId: browserAgentDownload
      security: []
      tags: [lan-files]
      description: Accepts only scoped browser_download or share_download tickets in the form body. Browser tickets require the signed Origin. Tickets in URLs and ordinary account bearer tokens are rejected. File bytes are streamed from the Agent, never through the control service.
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required: [ticket]
              properties:
                ticket: { type: string }
      responses:
        '200': { description: Verified attachment bytes }
        '401': { description: Expired or invalid ticket }
        '403': { description: Wrong scope or Origin }
        '404': { description: File absent in the authorized account }
...
  /v1/devices/register:
    post:
      summary: Register or resume a device identity
      operationId: registerDevice
      tags: [devices]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterDeviceRequest'
      responses:
        '201':
          description: Independent device session
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthResponse' }
        '409':
          description: Peer identity belongs to another account
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
...
  /v1/devices/{device_id}:
    parameters:
      - name: device_id
        in: path
        required: true
        schema: { type: string, format: uuid }
    patch:
      summary: Rename an active device owned by the current account
      operationId: renameDevice
      tags: [devices]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string, minLength: 1, maxLength: 128 }
      responses:
        '204': { description: Device renamed }
        '404': { description: Device not found in this account }
    delete:
      summary: Deregister an account device and revoke its sessions
      operationId: deregisterDevice
      tags: [devices]
      responses:
        '204': { description: Device deregistered }
        '404':
          description: Device not found in this account
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /v1/devices/heartbeat:
    post:
      summary: Record authenticated device liveness
      operationId: deviceHeartbeat
      tags: [devices]
      responses:
        '204': { description: Liveness recorded }
        '401':
          description: Session is not active
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
...
  /v1/catalog/files/register:
    post:
      summary: Register a verified Agent-local file and replica
      operationId: registerCatalogFile
      tags: [catalog]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RegisterCatalogFileRequest' }
      responses:
        '200':
          description: Registered or existing unified file
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UnifiedFile' }
        '409':
          description: Active account filename conflict
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /v1/catalog/lifecycle:
    post:
      summary: Apply a committed origin-Agent lifecycle mutation
      operationId: reportCatalogLifecycle
      tags: [catalog]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CatalogLifecycleReport' }
      responses:
        '200':
          description: Updated unified file
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UnifiedFile' }
        '404':
          description: File is not owned by the reporting Agent
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /v1/transfers:
    get:
      summary: List account transfer tasks
      operationId: listTransfers
      tags: [transfers]
      parameters:
        - { name: state, in: query, schema: { type: string } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 500 } }
      responses:
        '200':
          description: Transfer tasks
          content:
            application/json:
              schema:
                type: object
                required: [transfers]
                properties:
                  transfers: { type: array, items: { $ref: '#/components/schemas/TransferTask' } }
    post:
      summary: Schedule a verified file replica on another account device
      operationId: createTransfer
      tags: [transfers]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateTransferRequest' }
      responses:
        '201':
          description: Transfer task created or existing live task returned
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TransferTask' }
        '409':
          description: Target already contains the replica or no source is ready
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
...
components:
  securitySchemes:
    browserSession:
      type: apiKey
      in: cookie
      name: share_disk_session
      description: Account-resource endpoints also accept this cookie; browser writes require exact Origin and X-CSRF-Token. Device-only endpoints require native device credentials.
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    BrowserCSRF:
      name: X-CSRF-Token
      in: header
      required: true
      schema: { type: string, minLength: 1 }
...
  schemas:
    BrowserSession:
      type: object
      required: [account, csrf_token, expires_at]
      properties:
        account:
          type: object
          required: [id, account]
          properties:
            id: { type: string, format: uuid }
            account: { type: string }
            created_at: { type: string, format: date-time }
        csrf_token: { type: string }
        expires_at: { type: string, format: date-time }
...
    AuthResponse:
      type: object
      required:
        - access_token
        - refresh_token
        - expires_in
        - token_type
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        expires_in:
          type: integer
          description: Token expiration time in seconds
        token_type:
          type: string
          example: Bearer
        user_id: { type: string, format: uuid }
        device_id: { type: string, format: uuid }
...
    RegisterDeviceRequest:
      type: object
      additionalProperties: false
      required: [name, platform, peer_id]
      properties:
        name: { type: string, minLength: 1, maxLength: 128 }
        platform: { type: string, minLength: 1, maxLength: 32 }
        public_key: { type: string, contentEncoding: base64 }
        peer_id: { type: string, minLength: 1, maxLength: 255 }
...
    RegisterCatalogFileRequest:
      type: object
      additionalProperties: false
      required: [local_file_id, name, mime, size, sha256, endpoint]
      properties:
        local_file_id: { type: string, minLength: 1, maxLength: 64 }
        name: { type: string, minLength: 1, maxLength: 255 }
        mime: { type: string }
        size: { type: integer, minimum: 0 }
        sha256: { type: string, pattern: '^[0-9a-f]{64}$' }
        endpoint: { type: string, format: uri, maxLength: 2048 }
    CatalogLifecycleReport:
      type: object
      additionalProperties: false
      required: [operation_id, local_file_id, action]
      properties:
        operation_id: { type: string, minLength: 1, maxLength: 128 }
        local_file_id: { type: string, minLength: 1, maxLength: 64 }
        action: { type: string, enum: [rename, trash, restore, purge] }
        name: { type: string, minLength: 1, maxLength: 255 }
        purge_after: { type: [string, 'null'], format: date-time }
    UnifiedFile:
      type: object
      required: [id, folder_id, local_file_id, name, mime, size, sha256, status, version, origin_device_id, origin_device_name, origin_endpoint, replica_endpoint, content_local_file_id, available, replicas, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        folder_id: { type: string, format: uuid }
        local_file_id: { type: string }
        name: { type: string }
        mime: { type: string }
        size: { type: integer, minimum: 0 }
        sha256: { type: string, pattern: '^[0-9a-f]{64}$' }
        status: { type: string, enum: [active, trashed, purged] }
        version: { type: integer, minimum: 1 }
        origin_device_id: { type: string, format: uuid }
        origin_device_name: { type: string }
        origin_last_seen_at: { type: [string, 'null'], format: date-time }
        origin_endpoint: { type: string, format: uri }
        replica_endpoint: { type: string, format: uri }
        content_local_file_id: { type: string }
        available:
          type: boolean
          description: True when at least one verified replica device has a recent heartbeat.
        replicas:
          type: array
          items: { $ref: '#/components/schemas/UnifiedFileReplica' }
        deleted_at: { type: [string, 'null'], format: date-time }
        purge_after: { type: [string, 'null'], format: date-time }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    UnifiedFileReplica:
      type: object
      required: [device_id, device_name, platform, state, online, is_origin]
      properties:
        device_id: { type: string, format: uuid }
        device_name: { type: string }
        platform: { type: string }
        state: { type: string, enum: [pending, ready, missing, corrupt, deleting, deleted] }
        online: { type: boolean }
        is_origin: { type: boolean }
        last_seen_at: { type: [string, 'null'], format: date-time }
...
    CreateTransferRequest:
      type: object
      additionalProperties: false
      required: [file_id, target_device_id]
      properties:
        file_id: { type: string, format: uuid }
        target_device_id: { type: string, format: uuid }
        priority: { type: integer, minimum: -100, maximum: 100, default: 0 }
    TransferTask:
      type: object
      required: [id, user_id, object_id, target_device_id, reason, state, priority, attempt, version, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        user_id: { type: string, format: uuid }
        object_id: { type: string, format: uuid }
        target_device_id: { type: string, format: uuid }
        reason: { type: string, enum: [pull, push, replication, redundancy] }
        state: { type: string }
        priority: { type: integer }
        lease_owner: { type: [string, 'null'], format: uuid }
        lease_until: { type: [string, 'null'], format: date-time }
        attempt: { type: integer, minimum: 0 }
        version: { type: integer, minimum: 1 }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
...
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              description: Machine-readable error code
            message:
              type: string
              description: Human-readable error message
            request_id:
              type: string
              description: Request identifier for debugging
            details:
              type: object
              additionalProperties: true
              description: Additional error details
...